+++ src/PostMessageAPI.js
@@ -21,7 +21,11 @@ export function addLocizeSavedHandler(hnd) {
 let pendingMsgs = [];
 export function onAddedKey(lng, ns, key, value) {
   const msg = {
-    message: 'added', lng, ns, key, value
+    message: 'added',
+    lng,
+    ns,
+    key,
+    value,
   };
   if (source) {
     source.postMessage(msg, origin);
@@ -39,9 +43,7 @@ export const locizePlugin = {
 
     addLocizeSavedHandler((res) => {
       res.updated.forEach((item) => {
-        const {
-          lng, ns, key, data
-        } = item;
+        const { lng, ns, key, data } = item;
         i18n.addResource(lng, ns, key, data.value, { silent: true });
         i18n.emit('editorSaved');
       });
@@ -52,7 +54,11 @@ export const locizePlugin = {
         if (!isUpdate) onAddedKey(lng, ns, k, val);
       };
     }
-  }
+
+    i18next.on('languageChanged', (lng) => {
+      setEditorLng(lng);
+    });
+  },
 };
 
 function getI18next() {
@@ -72,9 +78,12 @@ if (typeof window !== 'undefined') {
       if (!source) {
         source = e.source;
         origin = e.origin;
-        handler = createClickHandler((payload) => {
-          source.postMessage({ message: 'clickedElement', payload }, origin);
-        });
+        handler = createClickHandler(
+          (payload) => {
+            source.postMessage({ message: 'clickedElement', payload }, origin);
+          },
+          { getI18next }
+        );
         // document.body.addEventListener('click', handler, true);
         // clickInterceptionEnabled = true;
       }
@@ -88,15 +97,19 @@ if (typeof window !== 'undefined') {
       });
       pendingMsgs = [];
     } else if (e.data.message === 'turnOn') {
-      if (scriptTurnedOff) return source.postMessage({ message: 'forcedOff' }, origin);
+      if (scriptTurnedOff)
+        return source.postMessage({ message: 'forcedOff' }, origin);
 
-      if (!clickInterceptionEnabled) window.document.body.addEventListener('click', handler, true);
+      if (!clickInterceptionEnabled)
+        window.document.body.addEventListener('click', handler, true);
       clickInterceptionEnabled = true;
       source.postMessage({ message: 'turnedOn' }, origin);
     } else if (e.data.message === 'turnOff') {
-      if (scriptTurnedOff) return source.postMessage({ message: 'forcedOff' }, origin);
+      if (scriptTurnedOff)
+        return source.postMessage({ message: 'forcedOff' }, origin);
 
-      if (clickInterceptionEnabled) window.document.body.removeEventListener('click', handler, true);
+      if (clickInterceptionEnabled)
+        window.document.body.removeEventListener('click', handler, true);
       clickInterceptionEnabled = false;
       source.postMessage({ message: 'turnedOff' }, origin);
     } else if (e.data.message === 'committed') {
@@ -110,7 +123,8 @@ if (typeof window !== 'undefined') {
 export function turnOn() {
   scriptTurnedOff = false;
 
-  if (!clickInterceptionEnabled) window.document.body.addEventListener('click', handler, true);
+  if (!clickInterceptionEnabled)
+    window.document.body.addEventListener('click', handler, true);
   clickInterceptionEnabled = true;
 
   if (source) source.postMessage({ message: 'turnedOn' }, origin);
@@ -120,10 +134,40 @@ export function turnOn() {
 export function turnOff() {
   scriptTurnedOff = true;
 
-  if (clickInterceptionEnabled) window.document.body.removeEventListener('click', handler, true);
+  if (clickInterceptionEnabled)
+    window.document.body.removeEventListener('click', handler, true);
   clickInterceptionEnabled = false;
 
   if (source) source.postMessage({ message: 'turnedOff' }, origin);
   if (source) source.postMessage({ message: 'forcedOff' }, origin);
   return scriptTurnedOff;
 }
+
+export function setEditorLng(lng) {
+  // console.warn('setLng', lng);
+  if (source) source.postMessage({ message: 'setLng', lng }, origin);
+}
+
+var oldHref = document.location.href;
+window.addEventListener('load', () => {
+  var bodyList = document.querySelector('body');
+
+  var observer = new MutationObserver(function (mutations) {
+    mutations.forEach(function (mutation) {
+      if (oldHref != document.location.href) {
+        // console.warn('url changed', oldHref, document.location.href);
+        oldHref = document.location.href;
+
+        if (source)
+          source.postMessage({ message: 'hrefChanged', href: oldHref }, origin);
+      }
+    });
+  });
+
+  var config = {
+    childList: true,
+    subtree: true,
+  };
+
+  observer.observe(bodyList, config);
+});