+++ src/clickHandler.js
@@ -1,10 +1,12 @@
 /* eslint-disable import/prefer-default-export */
 import {
-  getClickedElement, getElementText, getElementI18nKey, getElementNamespace
+  getClickedElement,
+  getElementText,
+  getElementI18nKey,
+  getElementNamespace,
 } from './utils';
 
-export function createClickHandler(cb) {
-
+export function createClickHandler(cb, options) {
   // eslint-disable-next-line consistent-return
   const handler = (e) => {
     const el = getClickedElement(e);
@@ -17,9 +19,7 @@ export function createClickHandler(cb) {
     const key = getElementI18nKey(el);
 
     const rectEl = el.getBoundingClientRect ? el : el.parentElement;
-    const {
-      top, left, width, height
-    } = rectEl.getBoundingClientRect();
+    const { top, left, width, height } = rectEl.getBoundingClientRect();
 
     const style = window.getComputedStyle(rectEl, null);
     const pT = parseFloat(style.getPropertyValue('padding-top'));
@@ -28,18 +28,25 @@ export function createClickHandler(cb) {
     const pL = parseFloat(style.getPropertyValue('padding-left'));
     const sizing = style.getPropertyValue('box-sizing');
 
+    function getFallbackNS() {
+      const i18next = options.getI18next();
+
+      if (i18next && i18next.options && i18next.options.isLocizify)
+        return i18next.options.defaultNS;
+    }
+
     cb({
       tagName: rectEl.tagName,
       text,
       key,
-      ns: getElementNamespace(el),
+      ns: getElementNamespace(el) || getFallbackNS(),
       box: {
         top,
         left,
         width: sizing === 'border-box' ? width : width - pR - pL,
-        height: sizing === 'border-box' ? height : height - pT - pB
+        height: sizing === 'border-box' ? height : height - pT - pB,
       },
-      style: style.cssText
+      style: style.cssText,
     });
   };